// src/core/types/entities/dma/approval/List.ts
export type List = {
  dailyScheduleId: number
  unitNumber: string
  equipmentModel: string
  brand: string
  smuDue: number
  workOrder: number
  psType: number
  dateService: string
  shift: string
  isActive: boolean
  startDate: string
  endDate: string
  createdOn: string
  createdBy: string
  changedOn: string
  changedBy: string
  status: null
}